[server] rename directories properly
