--- cmd/exposed-server/daemon/daemon.go
@@ -1,2 +1,0 @@
-// Package daemon is responsible for running the exposed-server in the background.
-package daemon
