+++ cmd/exposed-server/server/config/config_test.go
