+++ cmd/exposed-server/server/handlers/version.go
