+++ cmd/exposed-server/server/handlers/version_test.go
