+++ cmd/ingest-service/main.go
@@ -1,0 +1,6 @@
+// Package main for the ingest-service.
+package main
+
+func main() {
+
+}
